python_parser/requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.1
transformers>=4.41.2
torch>=2.3.0
httpx>=0.27.0
